refactor: Simplify calendar components by delegating to specialized modules
- Refactor GridComponents.kt and NavigationComponents.kt to act as facades, delegating to specialized header, grid, and navigation components for improved maintainability and organization.
- Remove legacy UI logic and streamline the structure for better clarity and consistency.
