+++ app/src/main/java/com/cocode/calendar/components/header/CalendarHeader.kt
@@ -1,0 +1,71 @@
+package com.cocode.calendar.components.header
+
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.cocode.calendar.CalColors
+import utils.Strings
+
+/**
+ * Displays a header row containing the days of the week.
+ *
+ * This composable function creates a horizontal row that shows the abbreviated names
+ * of the days of the week (Sun, Mon, Tue, etc.). The row is styled with a border
+ * and rounded corners at the top.
+ *
+ * The function uses [remember] to memorize the list of day names, preventing
+ * unnecessary recomposition.
+ *
+ * @see DayOfWeekBox
+ */
+@Composable
+fun WeekDaysHeader() {
+    val daysOfWeek = remember { Strings.Calendar.DAYS_OF_WEEK }
+
+    Row(
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically,
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 4.dp, vertical = 0.dp)
+            .border(
+                width = 1.dp,
+                color = CalColors.day_background,
+                RoundedCornerShape(10.dp, 10.dp, 0.dp, 0.dp)
+            )
+    ) {
+        daysOfWeek.forEach { day ->
+            DayOfWeekBox(day)
+        }
+    }
+}
+
+/**
+ * Displays a box containing the day of the week.
+ *
+ * This composable function creates a box with centered text representing a day of the week.
+ * The text color is different for weekdays and weekends.
+ *
+ * @param day The string representation of the day of the week (e.g., "Mon", "Tue").
+ */
+@Composable
+fun DayOfWeekBox(day: String) {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = Modifier
+            .width(55.dp)
+            .height(40.dp)
+    ) {
+        val color = if(day != Strings.Calendar.SUN && day != Strings.Calendar.SAT) CalColors.weekday_text else CalColors.weekend_text
+        Text(
+            text = day,
+            color = color
+        )
+    }
+}
