+++ app/src/main/java/com/cocode/calendar/components/pickers/CalendarPickerDialogs.kt
@@ -1,0 +1,13 @@
+package com.cocode.calendar.components.pickers
+
+/**
+ * CalendarPickerDialogs.kt - Legacy facade file.
+ *
+ * This file previously contained picker dialog implementations.
+ * The implementations have been moved to separate files:
+ * - MonthPickerDialog.kt for month selection
+ * - YearPickerDialog.kt for year selection
+ *
+ * The functions are now re-exported through NavigationComponents.kt
+ * for backward compatibility.
+ */
